refactor: Comments 컴포넌트 분리
