+++ client/src/components/Creator/Profile/Profile.styles.tsx
@@ -1,5 +1,5 @@
 import styled from 'styled-components';
-import Container from '../@atom/Container/Container';
+import Container from '../../@atom/Container/Container';
 
 export const StyledProfileContainer = styled(Container)``;
