+++ client/src/components/Creator/Profile/Profile.tsx
@@ -1,10 +1,10 @@
 import { FC, HTMLAttributes } from 'react';
 
-import Title from '../@atom/SubTitle/SubTitle';
+import Title from '../../@atom/SubTitle/SubTitle';
 import { ProfileImg, StyledProfileContainer } from './Profile.styles';
 
-import DefaultUserProfile from '../../assets/images/defualtUserProfile.png';
-import Anchor from '../@atom/Anchor/Anchor';
+import DefaultUserProfile from '../../../assets/images/defualtUserProfile.png';
+import Anchor from '../../@atom/Anchor/Anchor';
 
 const Profile: FC<HTMLAttributes<HTMLElement>> = () => {
   return (
